# Swagger/swagger.yaml
swagger: "2.0"
info:
  description: "API RESTFUL."
  version: "1.0.0"
  title: "Delilah Resto"
  termsOfService: "http://swagger.io/terms/"
  contact: 
    email: "hola@acamica.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "DelilahResto.com"
...
schemes:
- "https"
- "http"
paths:
...
  /users/delete/{Id}:
    delete:
      tags:
      - "user"
      summary: "Delete a User"
      description: "This can only be done by a admin"
      operationId: "deleteUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "Token"
        description: "Authorization: Bearer [token]"
        in: "header"
        required: true
        type: "string"
      - name: "Id"
        in: "path"
        description: "ID of product to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedUserError'
      security:
        - bearerAdminAuth: []
  /users/populateTables:
    get:
      tags:
      - "user"
      summary: "Populate Tables with random data"
      description: "This can be done without token"
      operationId: "populateTables"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: "The Tables have been successfully populated"
        500:
          description: "Internal Error"
...
  /products/{Id}:
    get:
      tags:
      - "product"
      summary: "Find product by ID"
      description: "Returns a single product. This can only be done with a Token, no matter if you are a user or a admin"
      operationId: "getProductById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "Token"
        description: "Authorization: Bearer [token]"
        in: "header"
        required: true
        type: "string"
      - name: "Id"
        in: "path"
        description: "ID of product to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Product"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Product not found"
        403:
          $ref: '#/responses/UnauthorizedBothError'
      security:
        - bearerBothAuth: []
    put:
      tags:
      - "product"
      summary: "Updates a product in the store. This can only be done by a admin"
      description: ""
      operationId: "updateProductWithForm"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "Token"
        description: "Authorization: Bearer [token]"
        in: "header"
        required: true
        type: "string"
      - name: "Id"
        in: "path"
        description: "/products/:id"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Product object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedUserError'
      security:
        - bearerAdminAuth: []
    delete:
      tags:
      - "product"
      summary: "Deletes a product. This can only be done by a admin"
      description: ""
      operationId: "deleteProduct"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "Token"
        description: "Authorization: Bearer [token]"
        in: "header"
        required: true
        type: "string"
      - name: "Id"
        in: "path"
        description: "Product id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedUserError'
      security:
        - bearerAdminAuth: []
...
securityDefinitions:
  bearerAdminAuth:
    type: apiKey
    in: header
    name: X-API-Key
  bearerUserAuth:
    type: apiKey
    in: header
    name: X-API-Key
  bearerBothAuth:
    type: apiKey
    in: header
    name: X-API-Key
responses:
...
  UnauthorizedUserError:
    description: Users are not allowed to make this request
    headers:
      Authorization:
        description: Bearer [token]
        type: string
  UnauthorizedBothError:
    description: Invalid Token
    headers:
      Authorization:
        description: Bearer [token]
        type: string
definitions:
...
  Product:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      description:
        type: "string"
      price:
        type: "integer"
      img:
        type: "string"
    xml:
      name: "Pet"
...